% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{db_connector}
\alias{db_connector}
\title{Creates function that checks for an existing DB connection and creates/open a new one if its not found open}
\usage{
db_connector(con_name, def_driver, checker = con_checker(con_name =
  con_name), def_args = NULL, def_args_nse = TRUE)
}
\arguments{
\item{con_name}{name of the connection}

\item{def_driver}{default DBI driver}

\item{checker}{function that can check the connection (returns TRUE or FALSE)}

\item{def_args}{list of default arguments (such as connection parameters, user name etc...) - see details}

\item{def_args_nse}{if TRUE (default) arguments in \code{def_args} get evaluated every time the new connectorfunction is called. This prevents situations when e.g. connector asks for a password and gets incorrect passwrod. Connectors created by \code{db_connector()} evaluates default arguments for the first time it is called and stores them (including the incorrect password) in the closure environment of the connector function and every new call of the connector no longer asks for password.}
}
\description{
Creates function that checks for an existing DB connection and creates/open a new one if its not found open
}
