% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{dbcon_env_exists}
\alias{dbcon_env_exists}
\title{Check whether .dbcon environment (for storing connections) exists}
\usage{
dbcon_env_exists()
}
\description{
Check whether .dbcon environment (for storing connections) exists
}
