% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R
\name{odbc_driver_lookup}
\alias{odbc_driver_lookup}
\title{Looks up best available odbc driver from given character vector of ODBC driver names.}
\usage{
odbc_driver_lookup(pref_drivers, pattern = "")
}
\arguments{
\item{pattern}{}
}
\description{
Looks up best available odbc driver from given character vector of ODBC driver names.
}
